+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Change Log
 5.1.1
 =====
 * Minor UI fixes to the Ingestion page.
+* Changed to NOT get Auth0 info from Portal; see foursight_core/react/api/auth0_config.py/PULL_AUTH0_INFO_FROM_PORTAL.
 
 
 5.1.0